feat: Improve linting configs and update husky setup
This commit expands and organises the config groups for linting, adding new groups for JSON and markdown files. It also adds logic to check for the existence of husky and creates Git hooks if husky is available. This serves to enhance the linting process, ensuring more comprehensive and automated checks. It also switches Eslint config file from '.eslintrc.cjs' to '.eslintrc.js' and includes additional file extensions (.json, .yaml, .yml) in the lint scope.

Additionally, the commit adds checks to ensure parallelism is disabled and debug mode is on in lint-staged. This aids in debugging any issues during linting and ensures stability. Furthermore, the commit adds 'secretlint-rule-preset-recommend' and 'secretlint' for detecting secrets in the code, providing additional security.

In terms of structural changes, related functions and constants have been grouped together and organised into utility modules, facilitating better readability and maintenance. Packages related to Windows have been removed as they are unnecessary here.

Signed-off-by: prisis <d.bannert@anolilab.de>