feat: Updated prepare-commit-msg script for Commitizen support
Added conditions to the 'prepare-commit-msg' Git hook script to check for presence of Commitizen and if found, to use it for commit message creation. This change facilitates adherence to conventional commits standard when Commitizen is part of the project.

Added also an escape condition for commit.cmd file and created this file to run Commitizen process interactively on Windows. For cross-platform compatibility, wrapped this section with OS check.

Lastly, added a disable comment for the SonarJS cognitive complexity rule to the `writeHuskyFiles` function due to complexity introduced by various conditional checks.

Signed-off-by: prisis <d.bannert@anolilab.de>