fix: Remove pre-push hook from postinstall script
Reduced the complexity of the postinstall script by removing the creation of the pre-push hook. This change was done to better focus on the core function of preparation for commit messages. Also, the condition for creating a prepare-commit-msg hook has been adjusted and it will now only be created if commitizen is present. Updated the README accordingly.

Signed-off-by: prisis <d.bannert@anolilab.de>