chore: Add semantic-release-npm-deprecate to deprecate old versions
In order to direct developers to use the latest versions of our packages, the semantic-release-npm-deprecate plugin has been added to our semantic-release preset. This plugin deprecates older versions of packages on npm and warns the user with a message to upgrade to a specified version. The plugin setup has also been added in the .releaserc.json files of individual packages where needed. Simultaneously, this plugin has been listed as a devDependency in the package.json files of several packages.

Signed-off-by: prisis <d.bannert@anolilab.de>