chore: Reorganize Prettier config and update README examples
This commit updates the `.prettierrc.js` configuration file to improve the readability of the options. Ordering of Prettier configuration options is now more logical. In addition to this, examples provided in README files of multiple packages (`prettier-config`, `browserslist-config-anolilab`, and `babel-preset`) are updated for more accuracy and to use module.exports. The changes make it easier for contributors to understand the configurations used in the project.

Signed-off-by: prisis <d.bannert@anolilab.de>