fix: Update file matching pattern to include nested directories
The file matching pattern in configurations for eslint, typescript, json and stylesheets have been updated from `*` to `**/*`. This change ensures that files in nested directories will also be matched and consequently, linted or formatted as needed. This was necessary as the former pattern was only scanning files in root and not the subdirectories.

Signed-off-by: prisis <d.bannert@anolilab.de>