fix: Refactor process calls and improve typings in babel-preset
Updated the postinstall.ts and index.ts files to import exit and env from the process. This change simplifies the handling of the process environment and process exits. The ESLint rules which were disabled due to process calls are also removed, enhancing code cleanliness. Additionally, interface was used instead of type for type definitions to improve consistency. The targets attribute in Options now aligns with the "targets" property in Babel's TransformOptions. The preset function return type is now denoted with TransformOptions to improve type safety.

Also updated packages in package.json to their latest versions, ensuring the package is up-to-date with the latest dependencies.

The above

Signed-off-by: prisis <d.bannert@anolilab.de>