fix: Refactor code to use type annotations and Record utility type
This change replaces object type definitions with Record utility type for more concise and standard type definitions in various files. This change also converts a type alias in eslint related files to an interface for better semantics and autocompletion. The dynamic type assertion in typescript and eslint related files is also refactored to improve readability. A types.d.ts file is newly created for centralizing commonly used types throughout the code base. Finally, the package.json file's dependencies version are updated to more recent ones to stay up to date. This commit brings clearer and standardized typings and paves the way for better type checking and autocompletion in the future.

Signed-off-by: prisis <d.bannert@anolilab.de>