fix: Update eslint configuration and plugin versions
Updated several eslint plugin versions in the package.json for eslint-config. This was carried out to ensure compatibility with latest versions and to incorporate the latest improvements/fixes from these libraries. In addition, the 'vitest/valid-expect' rule in vitest.ts plugin configuration was adjusted to enforce a minimum of 1 argument.

Signed-off-by: prisis <d.bannert@anolilab.de>