fix: Split 'testing-library' into 'testing-library-dom' and 'testing-library-react'
The 'testing-library' was split into 'testing-library-dom' and 'testing-library-react' to better organize and handle different configurations. With this change, it's now clearer which config and dependencies are used and it should be easier to adjust or extend specific behavior in the future. The renaming of files and changes in 'config.ts' reflect this separation.

Signed-off-by: prisis <d.bannert@anolilab.de>