fix: Remove 'n/file-extension-in-import' rule from overrides
The 'n/file-extension-in-import' rule has been removed from the 'overrides' section and configured as 'off'. This was done due to the rule being buggy as reported in the issue https://github.com/eslint-community/eslint-plugin-n/issues/21. Some code comments were also added to describe certain rules. The purpose of this change is to prevent undesired ESLint warnings.

Signed-off-by: prisis <d.bannert@anolilab.de>