feat: Add eslint-define-config and bin command for auto-generation of config
This change introduces the `eslint-define-config` package and an associated bin command that automatically creates an eslint configuration file. This allows the user to generate the configuration file without having to manually create one. It also restructures some of the code for better organization and readability, separates the `write-eslint-rc` and `write-eslint-ignore` functions into separate files, and updates the README accordingly. This change improves ease of setup and usage for the end-user. This commit introduces a new script, 'generate-eslint-config.js', which aids a user in
generating the '.eslintrc.js' file if for some reason the postinstall script doesn't run.

Signed-off-by: prisis <d.bannert@anolilab.de>