fix: update dependencies and improve installation instructions
Updated the version of various dependencies in the package.json file to use the latest versions. This was done to ensure the codebase is up-to-date and making use of the latest improvements and bug fixes of the dependencies.

Also, the installation instructions in the README.md file were amended to provide clearer information on setting up the 'eslint-plugin-import', as it had caused some confusion.

Furthermore, added 'typescript: true' in import.ts configuration file to enable TypeScript support.

Signed-off-by: prisis <d.bannert@anolilab.de>