fix: Downgrade semantic-release and conventional-changelog packages for compatibility
This commit includes a downgrade of the "semantic-release" and "conventional-changelog-conventionalcommits" packages
to ensure compatibility with other dependencies in the project. Changes were also made to ".ncurc.json" and
"renovate.json5" to reflect these downgrades to avoid breaking changes introduced in newer versions. The "pnpm-lock.yaml"
has been updated to reflect the new package versions.

Signed-off-by: prisis <d.bannert@anolilab.de>