fix: Refactor ESLint configurations for improved readability and consistency
This commit conducts a significant restructuring of the ESLint configurations. Numerous rule descriptions were modified for better clarity and consistency with the currently applied guidelines. Some of the rules were probably disabled or altered based on the current requirements. Also, introduced the usage of 'packageIsTypeModule' property from '@anolilab/package-json-utils' to manage the behavior of several rules depending on whether the module is a TypeScript module or not. This would help in making the linting process more adaptable to the type of module used.

Signed-off-by: prisis <d.bannert@anolilab.de>