fix: Update eslint-config including tests
This commit includes significant updates to the eslint-config package, primarily focusing on adding tests, updating dependencies, and refining configurations. It also includes updates for related configurations and dependencies within other files such as the package.json. The eslint-config updates include the addition of 'test' and 'test:coverage' scripts to offer more robust testing capabilities. Many dependencies within the package have been bumped to newer versions. Notably the "@eslint/js" package is updated to "^8.50.0", alongside updates to various rules and configurations. A new suite of tests has also been added to verify the old config installations and operation. This commit is part of an ongoing effort to ensure that our packages remain up-to-date, robust and maintainable. It bolsters the testing capabilities of the eslint-config package, helping to prevent regressions and ensure stability as new changes are introduced. This change will also refine and improve the linting rules and configurations used throughout the project.

Signed-off-by: prisis <d.bannert@anolilab.de>