fix: Handle failed promise in removeIgnoredFiles function
Error handling is added in removeIgnoredFiles function in ESLint config to better address instances when a promise fails. Instead of just directly awaiting on the promises sent by the `eslint.isPathIgnored()` function, the `Promise.allSettled()` method is now being used. This allows us to wait until all promises are settled, then check each promise's status. If a promise fails, an appropriate error message is output, and the process is exited. This prevents the promise-rejection-unhandled exception and ensures more robust functioning of the lint-stage-config.

Signed-off-by: prisis <d.bannert@anolilab.de>