fix: Update import/no-useless-path-segments rule and eslint dependency versions
Updated the rule for no-useless-path-segments in import.ts, enabling commonjs option instead of disabling it. This change was made in alignment with the updated version of eslint dependency.

Also, upgraded the versions of eslint-import resolver, eslint-patch, and eslint-plugin-security in package.json. This change resolves potential security risks and ensures compatibility with the new rules.

Signed-off-by: prisis <d.bannert@anolilab.de>